requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pygame==2.5.2
PyOpenGL==3.1.7
PyWavefront==1.3.3
requests==2.31.0
urllib3==2.1.0
